minor fixes and refactoring

+++ src/cuda/cudnn.jl
@@ -49,8 +49,8 @@ macro cudnncall(fv, argtypes, args...)
   f = eval(fv)
   quote
     _curet = ccall( ($(Meta.quot(f)), $libcudnn), Cint, $argtypes, $(args...)  )
-    if round(Int64, _curet) != CUDNN_STATUS_SUCCESS
-      throw(CuDNNError(round(Int64, _curet)))
+    if round(Int, _curet) != CUDNN_STATUS_SUCCESS
+      throw(CuDNNError(round(Int, _curet)))
     end
   end
 end
